Create README.md
# TableroPuntos-CheckMat-Brazil
This software was created and used in jiujitsu events, both in Argentina and in Curitiba Brazil. It has 4 panels (1 controller + 1 point marker + 1 advertising controller + 1 advertising panel)
Windows Forms technology and C# language were used and the work was done individually.
It is designed to connect a computer with a TV via HDMI cable and set to "expand screen" mode.